+++ reliably_cli/__main__.py
@@ -5,6 +5,7 @@ import typer
 from .__version__ import __version__
 from .config.cli import cli as config_cli
 from .config.types import Settings
+from .library.cli import cli as library_cli
 from .log import console
 from .services.cli import cli as services_cli
 
@@ -12,6 +13,7 @@ cli = typer.Typer()
 cli.add_typer(config_cli, name="config")
 # cli.add_typer(agent_cli, name="agent")
 cli.add_typer(services_cli, name="service")
+cli.add_typer(library_cli, name="library")
 
 
 @cli.callback()