+++ reliably_cli/library/cli.py
@@ -1,0 +1,6 @@
+import typer
+
+from .starters import cli as starters_cli
+
+cli = typer.Typer(help="Manage your libraries")
+cli.add_typer(starters_cli, name="starter")